requirements.txt
streamlit
python-dotenv
sqlite3
google-generativeai
